increase size of userids (#101)
* up size of user ids

* up size of user ids

+++ sql/cw_requests.sql
@@ -9,7 +9,7 @@ CREATE TABLE /*_*/cw_requests (
   cw_status VARCHAR(16) DEFAULT NULL,
   cw_timestamp varchar(32) NOT NULL,
   cw_url VARCHAR(96) NOT NULL,
-  cw_user SMALLINT NOT NULL,
+  cw_user INT(10) NOT NULL,
   cw_custom VARCHAR(96) NOT NULL,
   cw_category VARCHAR(64) NOT NULL,
   cw_visibility TINYINT UNSIGNED NOT NULL DEFAULT '0'
